delete_cloudinary

+++ routes/department.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 const pako =require('pako');
 const Home = require('../models/Department')
-const {img}=require('../middleware/cloudinary')
+const {img,deleteImageByUrl}=require('../middleware/cloudinary')
 ///////////////////////////create education page document//////////////////////
 router.post('/departmentdata', async (req, res) => {
     let success = false;
@@ -79,7 +79,13 @@ router.delete('/deletedepartment/:id', async (req, res) => {
     const id = req.params.id;
     let success = false;
     try {
-        const finddata = await Home.findByIdAndDelete(id);
+        let finddata= await Home.findById(id);
+        if(!finddata){
+            return res.status(400).send('Not found');
+        }
+
+        await deleteImageByUrl(finddata.picture);
+         finddata = await Home.findByIdAndDelete(id);
         success = true;
         res.json({ finddata, success })
