+++ routes/ethics.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const Home = require('../models/Ethics')
-const {img}=require('../middleware/cloudinary')
+const {img,deleteImageByUrl}=require('../middleware/cloudinary')
 const pako =require('pako');
 ///////////////////////////create education page document//////////////////////
 router.post('/ethicsdata', async (req, res) => {
@@ -78,7 +78,13 @@ router.delete('/deleteethics/:id', async (req, res) => {
     const id = req.params.id;
     let success = false;
     try {
-        const finddata = await Home.findByIdAndDelete(id);
+        let finddata= await Home.findById(id);
+        if(!finddata){
+            return res.status(400).send('Not found');
+        }
+
+        await deleteImageByUrl(finddata.picture);
+         finddata = await Home.findByIdAndDelete(id);
         success = true;
         res.json({ finddata, success })
