+++ routes/love.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = express.Router();
 const Home = require('../models/Love')
-const {img}=require('../middleware/cloudinary')
+const {img,deleteImageByUrl}=require('../middleware/cloudinary')
 const pako =require('pako');
 ///////////////////////////create education page document//////////////////////
 router.post('/lovedata', async (req, res) => {
@@ -74,12 +74,18 @@ router.get('/singlelove/:id',async (req,res)=>{
         res.json({ error, success })
     }
 })
-/////////////////////////////////delete love data//////////////////////
+/////////////////////////////////delete love data///////////////////////////////////////
 router.delete('/deletelove/:id', async (req, res) => {
     const id = req.params.id;
     let success = false;
     try {
-        const finddata = await Home.findByIdAndDelete(id);
+        let finddata= await Home.findById(id);
+        if(!finddata){
+            return res.status(400).send('Not found');
+        }
+
+        await deleteImageByUrl(finddata.picture);
+         finddata = await Home.findByIdAndDelete(id);
         success = true;
         res.json({ finddata, success })
